chore: create unit test for sdk client cookie helper
