+++ apps/web/sdk.client.ts
@@ -1,25 +1,6 @@
 import axios, { AxiosError } from 'axios';
 import { SdkHttpError } from '@vue-storefront/sdk';
-
-const VSL_LOCALE_COOKIE = 'vsf-locale';
-
-const updateVsfLocale = (cookie: string, locale: string): string => {
-  let cookieExists = false;
-  const cookies = (cookie ?? '').split(';');
-  const cookiesArray = cookies.map((cookie) => {
-    if (cookie.trim().includes(VSL_LOCALE_COOKIE)) {
-      cookieExists = true;
-      return `${VSL_LOCALE_COOKIE}=${locale}`;
-    }
-    return cookie;
-  });
-
-  if (!cookieExists) {
-    cookiesArray.push(`${VSL_LOCALE_COOKIE}=${locale}`);
-  }
-
-  return cookiesArray.join('; ');
-};
+import { updateVsfLocale } from './utils/sdkClientHelper';
 
 export const httpClient = async (url: any, params: any, config: any) => {
   try {
